+++ app/clients/dropbox/authenticate/switch_account.js
