--- app/clients/dropbox/dashboard/middleware/index.js
@@ -1,5 +1,0 @@
-var middleware = {};
-
-middleware.load_dropbox_account = require('./load_dropbox_account');
-
-module.exports = middleware;
