+++ app/clients/dropbox/index.js
@@ -5,6 +5,6 @@ module.exports = {
   disconnect: require("./disconnect"),
   remove: require("./remove"),
   write: require("./write"),
-  site_routes: require("./site"),
-  dashboard_routes: require("./dashboard")
+  site_routes: require("./routes").site,
+  dashboard_routes: require("./routes").dashboard
 };
