+++ app/clients/dropbox/views/permission.html
